Revert "Use bootstrap script in main for table creation"
This reverts commit 20fccebce2cc6b3c316fbc0e8326f3a77fc2c743.

+++ alembic/versions/e4fc04d1a442_add_annotations_to_tables.py
@@ -31,6 +31,14 @@ def upgrade() -> None:
     table. It includes a server-side default of an empty JSON object ('{}') to ensure
     that existing rows get a non-null default value.
     """
+    bind = op.get_bind()
+    sess = Session(bind=bind)
+    inspector = sa.inspect(bind)
+
+    if not inspector.has_table("gateways"):
+        print("Fresh database detected. Skipping migration.")
+        return
+
     op.add_column('tools', sa.Column('annotations', sa.JSON(), server_default=sa.text("'{}'"), nullable=False))
 
 
@@ -40,5 +48,13 @@ def downgrade() -> None:
 
     This function provides a way to undo the migration, safely removing the
     'annotations' column from the 'tool' table.
-    """ 
+    """
+    bind = op.get_bind()
+    sess = Session(bind=bind)
+    inspector = sa.inspect(bind)
+    
+    if not inspector.has_table("gateways"):
+        print("Fresh database detected. Skipping migration.")
+        return
+    
     op.drop_column('tools', 'annotations')
